requirements.txt
python-dotenv
rich
google-generativeai

